Fix format specifier warning at src/pet/petmem.c:458:47
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@36549 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/pet/petmem.c
@@ -455,7 +455,7 @@ static int efe0_dump(void)
         int sh = dongle6702.shift[i];
         int lm = leftmost[i];
 
-        mon_out("%d %3d: $%02x  %%", i, mask, sh);
+        mon_out("%d %3d: $%02x  %%", i, mask, (unsigned int)sh);
 
         for (j = 7, maskj = 1 << j; j >= 0; j--, maskj >>= 1) {
             if (maskj > lm) {
